+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/mapper/SessionEntityMapperTest.java
@@ -12,7 +12,6 @@ import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSession;
 import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerVote;
 import de.codescape.jira.plugins.scrumpoker.model.AllowRevealDeck;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
-import de.codescape.jira.plugins.scrumpoker.model.SpecialCards;
 import de.codescape.jira.plugins.scrumpoker.rest.entities.SessionEntity;
 import de.codescape.jira.plugins.scrumpoker.rest.entities.VoteEntity;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerCardService;
@@ -83,8 +82,6 @@ public class SessionEntityMapperTest {
         when(userSpecificDateTimeFormatter.withStyle(ArgumentMatchers.any(DateTimeStyle.class)))
             .thenReturn(userSpecificDateTimeFormatter);
         when(dateTimeFormatter.forLoggedInUser()).thenReturn(userSpecificDateTimeFormatter);
-        when(scrumPokerCardService.getCardSet()).thenReturn(
-            Arrays.asList(QUESTION_MARK, COFFEE_CARD, "0", "1", "2", "3", "5", "8", "13", "20", "40", "100"));
     }
 
     @Test
@@ -141,6 +138,8 @@ public class SessionEntityMapperTest {
             scrumPokerVote("5"),
             scrumPokerVote("13")};
         ScrumPokerSession scrumPokerSession = scrumPokerSession(scrumPokerVotes, true);
+        when(scrumPokerCardService.getCardSet(ArgumentMatchers.any(ScrumPokerSession.class))).thenReturn(
+            Arrays.asList(QUESTION_MARK, COFFEE_CARD, "0", "1", "2", "3", "5", "8", "13", "20", "40", "100"));
 
         SessionEntity sessionEntity = sessionEntityMapper.build(scrumPokerSession, CURRENT_USER);
         assertThat(sessionEntity.getBoundedVotes().size(), is(equalTo(3)));